Introduce the new songbook format.
This release now uses JSON coded data to represent a songbook. Songbook are stored in .sb files. The generation of the corresponding .tex file is done by the songbook.py program and uses templates stored in templates/. The default template is songbook.tmpl. It produces songbooks similar to the previous version of the songbook software suite.