Add repeatchords option to the crepbook document class
The repeatchords option enables chords printing in repeated chords
section. The new repeatedchords environment is used by song authors
to declare that chords enclosed in the environment are only the
repeated version of the chords given earlier. This environment allows
users who prefer to repeat chords on every verse and chorus to write
them down, and in conjunction with the repeatchords option to have
songbooks that displays the repeated chords.