major performance improvement when building a songbook
make songbook.pdf :
 - before patch: 5mn23s
 - after patch: 14s

The issue was about LaTeX \graphicspath macro that was used for
covers.  Since covers are stored in each artist's directory,
\graohicspath was containing a large number of different paths which
resulted in quite some time to evaluate when using \includegraphics.

This small patch copies all covers in a single temporary directory
before building a songbook. Note that symlinks are quite tricky to use
with pdflatex, esp. with MacOs/Windows plateforms so a hard copy has
been preferred.